Move entity Actions to actions.entities
