+++ Client/src/main/java/de/turtle_exception/client/api/TurtleClient.java
@@ -13,9 +13,9 @@ import de.turtle_exception.client.api.entities.messages.SyncMessage;
 import de.turtle_exception.client.api.event.EventListener;
 import de.turtle_exception.client.api.event.EventManager;
 import de.turtle_exception.client.api.request.Action;
-import de.turtle_exception.client.api.request.GroupAction;
-import de.turtle_exception.client.api.request.TicketAction;
-import de.turtle_exception.client.api.request.UserAction;
+import de.turtle_exception.client.api.request.entities.GroupAction;
+import de.turtle_exception.client.api.request.entities.TicketAction;
+import de.turtle_exception.client.api.request.entities.UserAction;
 import de.turtle_exception.client.internal.NetworkAdapter;
 import de.turtle_exception.client.internal.Provider;
 import de.turtle_exception.client.internal.net.NetClient;
