+++ Client/src/main/java/de/turtle_exception/client/api/request/entities/TicketAction.java
@@ -1,9 +1,10 @@
-package de.turtle_exception.client.api.request;
+package de.turtle_exception.client.api.request.entities;
 
 import de.turtle_exception.client.api.TurtleClient;
 import de.turtle_exception.client.api.entities.Ticket;
 import de.turtle_exception.client.api.entities.User;
 import de.turtle_exception.client.api.entities.attributes.TicketState;
+import de.turtle_exception.client.api.request.Action;
 import net.dv8tion.jda.api.entities.channel.middleman.GuildMessageChannel;
 import org.jetbrains.annotations.NotNull;
