+++ Client/src/main/java/de/turtle_exception/client/internal/TurtleClientImpl.java
@@ -3,26 +3,25 @@ package de.turtle_exception.client.internal;
 import com.google.gson.JsonObject;
 import de.turtle_exception.client.api.TurtleClient;
 import de.turtle_exception.client.api.entities.*;
-import de.turtle_exception.client.api.entities.containers.ITurtleContainer;
 import de.turtle_exception.client.api.entities.messages.DiscordChannel;
 import de.turtle_exception.client.api.entities.messages.MinecraftChannel;
 import de.turtle_exception.client.api.entities.messages.SyncChannel;
 import de.turtle_exception.client.api.entities.messages.SyncMessage;
 import de.turtle_exception.client.api.event.EventManager;
 import de.turtle_exception.client.api.request.Action;
-import de.turtle_exception.client.api.request.GroupAction;
-import de.turtle_exception.client.api.request.TicketAction;
-import de.turtle_exception.client.api.request.UserAction;
+import de.turtle_exception.client.api.request.entities.GroupAction;
+import de.turtle_exception.client.api.request.entities.TicketAction;
+import de.turtle_exception.client.api.request.entities.UserAction;
 import de.turtle_exception.client.internal.data.ResourceBuilder;
 import de.turtle_exception.client.internal.data.annotations.Keys;
 import de.turtle_exception.client.internal.entities.*;
 import de.turtle_exception.client.internal.event.UpdateHelper;
 import de.turtle_exception.client.internal.net.NetClient;
 import de.turtle_exception.client.internal.net.NetworkProvider;
-import de.turtle_exception.client.internal.request.actions.GroupActionImpl;
+import de.turtle_exception.client.internal.request.actions.entities.GroupActionImpl;
 import de.turtle_exception.client.internal.request.actions.SimpleAction;
-import de.turtle_exception.client.internal.request.actions.TicketActionImpl;
-import de.turtle_exception.client.internal.request.actions.UserActionImpl;
+import de.turtle_exception.client.internal.request.actions.entities.TicketActionImpl;
+import de.turtle_exception.client.internal.request.actions.entities.UserActionImpl;
 import de.turtle_exception.client.internal.util.TurtleSet;
 import de.turtle_exception.client.internal.util.version.IllegalVersionException;
 import de.turtle_exception.client.internal.util.version.Version;
